rename publicprofilestack to userprofilestack, move bottom tab into main stack
