+++ src/navigation/EventsStack.tsx
@@ -10,7 +10,6 @@ import SetGeneralEventDetails from "../screens/events/SetGeneralEventDetails";
 import SetSpecificEventDetails from "../screens/events/SetSpecificEventDetails";
 import FinalizeEvent from "../screens/events/FinalizeEvent";
 import SetLocationEventDetails from "../screens/events/SetLocationEventDetails";
-import QRCodeScanningScreen from "../screens/events/QRCodeScanningScreen";
 import EventVerification from "../screens/events/EventVerification";
 import PublicProfileScreen from "../screens/PublicProfile";
 
@@ -19,10 +18,9 @@ const EventsStack = () => {
     return (
         <Stack.Navigator screenOptions={{ headerShown: false }}>
             <Stack.Screen name="EventsScreen" component={Events} />
-            <Stack.Screen name="UpdateEvent" component={UpdateEvent} />
             <Stack.Screen name="EventInfo" component={EventInfo} />
+            <Stack.Screen name="UpdateEvent" component={UpdateEvent} />
             <Stack.Screen name="QRCode" component={QRCodeManager} />
-            <Stack.Screen name="QRCodeScanningScreen" component={QRCodeScanningScreen} />
 
             {/* Event Creation Screens */}
             <Stack.Group>
