+++ src/screens/events/PersonalEventLog.tsx
@@ -8,9 +8,9 @@ import { UserContext } from '../../context/UserContext';
 import { auth } from '../../config/firebaseConfig';
 import { queryUserEventLogs } from '../../api/firebaseUtils';
 import { UserEventData } from '../../types/events';
-import { PublicProfileStackParams } from '../../types/navigation';
+import { UserProfileStackParams } from '../../types/navigation';
 
-const PersonalEventLog = ({ navigation }: NativeStackScreenProps<PublicProfileStackParams>) => {
+const PersonalEventLog = ({ navigation }: NativeStackScreenProps<UserProfileStackParams>) => {
   const userContext = useContext(UserContext);
   const { userInfo } = userContext!;
